requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
emoji==0.6.0
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.0
pytz==2020.1
six==1.15.0
tornado==6.0.4
tzlocal==2.1
